bot/requirements.txt
aiogram==2.14.3
psycopg2
fabric
paramiko
python-dotenv
